+++ migrations/2025-09-02-220159_add_external_integrations_settings/down.sql
@@ -1,0 +1,8 @@
+-- Remove email settings
+DELETE FROM system_settings WHERE category = 'email';
+
+-- Remove OAuth settings
+DELETE FROM system_settings WHERE category = 'oauth';
+
+-- Remove storage settings
+DELETE FROM system_settings WHERE category = 'storage';